+++ src/EventSpaceAlgebra.jl
@@ -1,6 +1,13 @@
 module EventSpaceAlgebra
-
+include("abstractspace.jl")
 include("timespacetypes.jl")
-export Longitude, Latitude, EventTime, Degree, JulianDay, AngularDistance, EpochDuration, ValueUnit, GeneralSpace, Distance
+include("algebra/sameunit.jl")
+include("algebra/comparisonop.jl")
+include("algebra/addsubtract.jl")
 
+export Longitude, Latitude, EventTime, Degree, JulianDay, AngularDistance, EpochDuration, ValueUnit
+
+# General interface for construction
+export Coordinate, Distance
+export get_unit, get_value, set_value
 end