+++ test/timespacetypes.jl
@@ -13,7 +13,7 @@
         EventTime(1, JulianDay), Distance(2.3, JulianDay)
     )
     @test EventSpaceAlgebra.sameunit(
-        Distance(1, Degree), GeneralSpace(Latitude, 1, Degree)
+        Distance(1, Degree), Coordinate(Latitude, 1, Degree)
     )
     @test_throws EventSpaceAlgebra.UnitMismatch EventSpaceAlgebra.sameunit(
         Distance(1, Degree), EventTime(1, JulianDay)
@@ -26,21 +26,89 @@
 
     # Test `sameunit` error
     @test_throws EventSpaceAlgebra.UnitMismatch EventSpaceAlgebra.sameunit(
-        GeneralSpace(EventTime, 1, JulianDay), GeneralSpace(Latitude, 2, Degree)
+        Coordinate(EventTime, 1, JulianDay), Coordinate(Latitude, 2, Degree)
     )
 
     # Test `isless`
     @test isless(
-        GeneralSpace(Latitude, 1, Degree), GeneralSpace(Latitude, 2, Degree)
+        Coordinate(Latitude, 1, Degree), Coordinate(Latitude, 2, Degree)
     )
     @test_throws MethodError isless(
-        GeneralSpace(Longitude, 1, Degree), GeneralSpace(Latitude, 1, Degree)
+        Coordinate(Longitude, 1, Degree), Coordinate(Latitude, 1, Degree)
     ) # Longitude and Latitude is not comparable in size even when they are of the same unit
 
     @test isapprox(
-        GeneralSpace(Longitude, 121.33, Degree) - GeneralSpace(Longitude, 110.0, Degree), Distance(11.33, Degree)
+        Coordinate(Longitude, 121.33, Degree) - Coordinate(Longitude, 110.0, Degree), Distance(11.33, Degree)
     )
-    # TODO:
-    # - Distance of the same unit is substractable/addable
-    # - T<:Spatial can be substracted by Y<:...
+    @test isapprox(
+        Coordinate(Longitude, 121.33, Degree) - Distance(110.0, Degree), Distance(11.33, Degree)
+    )
+    @test isapprox(
+        Distance(121.33, Degree) - Distance(110.0, Degree), Distance(11.33, Degree)
+    )
+
+    @test isapprox(
+        Coordinate(EventTime, 121.33, JulianDay) - Coordinate(EventTime, 110.0, JulianDay), Distance(11.33, JulianDay)
+    )
+    @test isapprox(
+        Coordinate(EventTime, 121.33, JulianDay) - Distance(110.0, JulianDay), Distance(11.33, JulianDay)
+    )
+    @test isapprox(
+        Distance(121.33, JulianDay) - Distance(110.0, JulianDay), Distance(11.33, JulianDay)
+    )
+
+    # test set_value
+    lg122 = Longitude(122, Degree)
+    @test isequal(set_value(lg122, 123), Longitude(123, Degree))
+
+end
+
+@testset "Test Addition" begin
+    # Coordinate of different type should not be subtractable
+    @test_throws EventSpaceAlgebra.CoordinateMismatch Coordinate(Longitude, 121.33, Degree) - Coordinate(Latitude, 22.3, Degree)
+    @test_throws EventSpaceAlgebra.CoordinateMismatch Coordinate(Longitude, 121.33, Degree) - Coordinate(EventTime, 22.3, JulianDay)
+
+    # Distance substracted by Coordinate is unreasonable
+    @test_throws EventSpaceAlgebra.CoordinateMismatch Distance(121.33, Degree) - Coordinate(Latitude, 22.3, Degree)
+    @test_throws EventSpaceAlgebra.CoordinateMismatch Distance(121.33, JulianDay) - Coordinate(EventTime, 22.3, JulianDay)
+
+    # test addition
+    @test_throws EventSpaceAlgebra.UnitMismatch EventSpaceAlgebra._create_add(
+        Longitude(1.1, Degree),
+        EventTime(1, JulianDay)
+    )
+    cs = [Longitude, Latitude, EventTime]
+    ut = [Degree, Degree, JulianDay]
+    for (constructor, unit) in zip(cs, ut)
+        @test_throws EventSpaceAlgebra.CoordinateMismatch constructor(1, unit) + constructor(1, unit)
+    end
+
+    c1s = [Longitude, Latitude, EventTime, Distance]
+    c2s = [Distance, Distance, Distance, Distance]
+    uts = [Degree, Degree, JulianDay, Degree]
+    for (c1, c2, unit) in zip(c1s, c2s, uts)
+        @test isequal(c1(1, unit) + c2(1, unit), c1(2, unit))
+    end
+end
+
+@testset "DataFrame grouping" begin
+    using DataFrames
+    df = DataFrame(
+        :Lat => repeat(22.1:0.1:22.3, inner=[4]),
+        :Lon => repeat(122.1:0.1:122.4, outer=[3]),
+        :Time => repeat(1.1:0.1:1.6, outer=[2]))
+
+    @test length(groupby(df, [:Lat])) == 3
+    @test length(groupby(df, [:Lon])) == 4
+    @test length(groupby(df, [:Time])) == 6
+
+    transform!(df, :Time => ByRow(x -> EventTime(x, JulianDay)); renamecols=false)
+    transform!(df, :Lat => ByRow(x -> Latitude(x, Degree)); renamecols=false)
+    transform!(df, :Lon => ByRow(x -> Longitude(x, Degree)); renamecols=false)
+    # hdf = nrow(df)
+
+    @test length(groupby(df, [:Lat])) == 3
+    @test length(groupby(df, [:Lon])) == 4
+    @test length(groupby(df, [:Time])) == 6
+
 end