+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
 ## v0.0.1
-
-  * Initiating the project.
+- Initiating the project.
+## v0.2.0
+- `Distance` and `Coordinate` in `AbstractSpace`
+- Methods of addition and subtraction for `Distance` and `Coordinate`